fix: check price when pseudo-matching OrderView (cherry-picked)
(cherry picked from commit 7e8aaee402c8b03924b5c24d9bbc20ce4e541b47)
