+++ x/liquidity/types/orderbook_test.go
@@ -15,97 +15,6 @@ func newOrder(dir amm.OrderDirection, price sdk.Dec, amt sdk.Int) amm.Order {
 	return amm.DefaultOrderer.Order(dir, price, amt)
 }
 
-func ExampleMakeOrderBookResponse() {
-	ob := amm.NewOrderBook(
-		newOrder(amm.Sell, utils.ParseDec("15.0"), sdk.NewInt(10000)),
-		newOrder(amm.Sell, utils.ParseDec("13.0"), sdk.NewInt(10000)),
-		newOrder(amm.Sell, utils.ParseDec("10.01"), sdk.NewInt(10000)),
-		newOrder(amm.Sell, utils.ParseDec("10.0"), sdk.NewInt(10000)),
-		newOrder(amm.Buy, utils.ParseDec("10.01"), sdk.NewInt(10000)),
-		newOrder(amm.Buy, utils.ParseDec("10.0"), sdk.NewInt(10000)),
-		newOrder(amm.Buy, utils.ParseDec("9.0"), sdk.NewInt(10000)),
-		newOrder(amm.Buy, utils.ParseDec("3.0"), sdk.NewInt(10000)),
-		newOrder(amm.Buy, utils.ParseDec("0.1"), sdk.NewInt(10000)),
-	)
-	ov := ob.MakeView()
-	ov.Match()
-	tickPrec := 1
-	resp := types.MakeOrderBookResponse(ov, tickPrec, 20)
-	basePrice, found := types.OrderBookBasePrice(ov, tickPrec)
-	if !found {
-		panic("base price not found")
-	}
-	types.PrintOrderBookResponse(resp, basePrice)
-
-	// Output:
-	// +------------------------------------------------------------------------+
-	// |              10000 |        15.000000000000000000 |                    |
-	// |              10000 |        13.000000000000000000 |                    |
-	// |              10000 |        11.000000000000000000 |                    |
-	// |------------------------------------------------------------------------|
-	// |                             10.000000000000000000                      |
-	// |------------------------------------------------------------------------|
-	// |                    |        10.000000000000000000 | 10000              |
-	// |                    |         9.000000000000000000 | 10000              |
-	// |                    |         3.000000000000000000 | 10000              |
-	// |                    |         0.000000000000000000 | 10000              |
-	// +------------------------------------------------------------------------+
-}
-
-func ExampleMakeOrderBookResponse_pool() {
-	pool1 := amm.NewBasicPool(sdk.NewInt(1050_000000), sdk.NewInt(1000_000000), sdk.Int{})
-	pool2, err := amm.CreateRangedPool(
-		sdk.NewInt(1000_000000), sdk.NewInt(1000_000000),
-		utils.ParseDec("0.1"), utils.ParseDec("5.0"), utils.ParseDec("0.95"))
-	if err != nil {
-		panic(err)
-	}
-
-	lastPrice := utils.ParseDec("1.0")
-	lowestPrice := lastPrice.Mul(utils.ParseDec("0.9"))
-	highestPrice := lastPrice.Mul(utils.ParseDec("1.1"))
-	ob := amm.NewOrderBook()
-	ob.AddOrder(amm.PoolOrders(pool1, amm.DefaultOrderer, lowestPrice, highestPrice, 4)...)
-	ob.AddOrder(amm.PoolOrders(pool2, amm.DefaultOrderer, lowestPrice, highestPrice, 4)...)
-
-	ov := ob.MakeView()
-	ov.Match()
-	tickPrec := 2
-	resp := types.MakeOrderBookResponse(ov, tickPrec, 20)
-	basePrice, found := types.OrderBookBasePrice(ov, tickPrec)
-	if !found {
-		panic("base price not found")
-	}
-	types.PrintOrderBookResponse(resp, basePrice)
-
-	// Output:
-	// +------------------------------------------------------------------------+
-	// |           12482141 |         1.100000000000000000 |                    |
-	// |           11488898 |         1.090000000000000000 |                    |
-	// |            8280038 |         1.080000000000000000 |                    |
-	// |           12100493 |         1.070000000000000000 |                    |
-	// |           12139873 |         1.060000000000000000 |                    |
-	// |            7381973 |         1.050000000000000000 |                    |
-	// |            6623054 |         1.040000000000000000 |                    |
-	// |            8863846 |         1.030000000000000000 |                    |
-	// |            7489932 |         1.020000000000000000 |                    |
-	// |            6572362 |         1.010000000000000000 |                    |
-	// |            8736758 |         1.000000000000000000 |                    |
-	// |------------------------------------------------------------------------|
-	// |                              0.990000000000000000                      |
-	// |------------------------------------------------------------------------|
-	// |                    |         0.980000000000000000 | 3406177            |
-	// |                    |         0.970000000000000000 | 5349384            |
-	// |                    |         0.960000000000000000 | 6256722            |
-	// |                    |         0.950000000000000000 | 4655698            |
-	// |                    |         0.940000000000000000 | 13393535           |
-	// |                    |         0.930000000000000000 | 13541701           |
-	// |                    |         0.920000000000000000 | 17097505           |
-	// |                    |         0.910000000000000000 | 14039668           |
-	// |                    |         0.900000000000000000 | 14822646           |
-	// +------------------------------------------------------------------------+
-}
-
 func TestMakeOrderBookResponse(t *testing.T) {
 	pool := amm.NewBasicPool(sdk.NewInt(862431695563), sdk.NewInt(37852851767), sdk.Int{})
 	lowestPrice := pool.Price().Mul(sdk.NewDecWithPrec(9, 1))
